project/requirements.txt
# requirements.txt

# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# Data
pandas==2.2.3
numpy==1.26.4

# ML
catboost==1.2.7
scikit-learn==1.6.1

# Config
pyyaml==6.0.2

# Serialization
joblib==1.4.2

# MLflow
mlflow==2.19.0

# Visualization
matplotlib==3.10.0
seaborn==0.13.2

# Testing
pytest==8.3.4
httpx==0.28.1
